backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-chess==1.999
opencv-python==4.9.0.80
ultralytics==8.1.0
yt-dlp==2023.12.30
numpy==1.26.3
pillow==10.2.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
matplotlib==3.8.2
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0